requirements/base.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==5.0.1
django-axes==6.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
identify==2.5.33
inflection==0.5.1
nodeenv==1.8.0
packaging==23.2
platformdirs==4.1.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
virtualenv==20.25.0
